DCMAW-5594: File restructure
